Bump fp-kontrakter 6.2.1 som inneholder package rename av fpwsporxy import
