+++ domenetjenester/simulering/src/test/java/no/nav/foreldrepenger/oppdrag/domenetjenester/simulering/StartSimuleringTjenesteFpWsProxyTest.java
@@ -23,23 +23,23 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeEndring;
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeEndringLinje;
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeFagområde;
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeKlassifik;
-import no.nav.foreldrepenger.kontrakter.simulering.request.KodeStatusLinje;
-import no.nav.foreldrepenger.kontrakter.simulering.request.LukketPeriode;
-import no.nav.foreldrepenger.kontrakter.simulering.request.Oppdrag110Dto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.OppdragskontrollDto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.Oppdragslinje150Dto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.Refusjonsinfo156Dto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.SatsDto;
-import no.nav.foreldrepenger.kontrakter.simulering.request.TypeSats;
-import no.nav.foreldrepenger.kontrakter.simulering.request.UtbetalingsgradDto;
-import no.nav.foreldrepenger.kontrakter.simulering.respons.BeregningDto;
-import no.nav.foreldrepenger.kontrakter.simulering.respons.BeregningStoppnivåDetaljerDto;
-import no.nav.foreldrepenger.kontrakter.simulering.respons.BeregningStoppnivåDto;
-import no.nav.foreldrepenger.kontrakter.simulering.respons.BeregningsPeriodeDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeEndring;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeEndringLinje;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeFagområde;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeKlassifik;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.KodeStatusLinje;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.LukketPeriode;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.Oppdrag110Dto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.OppdragskontrollDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.Oppdragslinje150Dto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.Refusjonsinfo156Dto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.SatsDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.TypeSats;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.request.UtbetalingsgradDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.respons.BeregningDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.respons.BeregningStoppnivåDetaljerDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.respons.BeregningStoppnivåDto;
+import no.nav.foreldrepenger.kontrakter.fpwsproxy.simulering.respons.BeregningsPeriodeDto;
 import no.nav.foreldrepenger.oppdrag.dbstoette.JpaExtension;
 import no.nav.foreldrepenger.oppdrag.domenetjenester.person.PersonTjeneste;
 import no.nav.foreldrepenger.oppdrag.domenetjenester.simulering.fpwsproxy.FpWsProxySimuleringKlient;
