+++ pom.xml
@@ -29,7 +29,7 @@
 
         <felles.version>4.2.16</felles.version>
         <!-- Kontrakter -->
-        <fp-kontrakter.version>6.1.31</fp-kontrakter.version>
+        <fp-kontrakter.version>6.2.1</fp-kontrakter.version>
     </properties>
 
     <dependencyManagement>
