+++ web/pom.xml
@@ -114,12 +114,6 @@
             <artifactId>weld-servlet-core</artifactId>
         </dependency>
 
-        <!-- For Status side -->
-        <dependency>
-            <groupId>org.apache.httpcomponents</groupId>
-            <artifactId>httpclient</artifactId>
-        </dependency>
-
         <!-- Server -->
         <dependency>
             <groupId>org.eclipse.jetty</groupId>
